feature: implement filtered rb
